move erosion utils from the `support.py` module to the `bank_erosion.utils.py`
