+++ src/dfastbe/bank_lines/bank_lines.py
@@ -16,7 +16,7 @@ from dfastbe.io.data_models import BaseSimulationData
 from dfastbe.io.config import ConfigFile
 from dfastbe.io.logger import log_text
 from dfastbe.bank_lines.utils import sort_connect_bank_lines, poly_to_line, tri_to_line
-from dfastbe.support import on_right_side
+from dfastbe.utils import on_right_side
 from dfastbe.io.logger import timed_logger
 
 MAX_RIVER_WIDTH = 1000
