+++ src/dfastbe/bank_lines/utils.py
@@ -1,7 +1,7 @@
 import numpy as np
 from shapely.geometry import LineString, MultiLineString, Point
 from shapely import line_merge
-from dfastbe.support import on_right_side
+from dfastbe.utils import on_right_side
 
 __all__ = ["sort_connect_bank_lines", "poly_to_line", "tri_to_line"]
